fbwifi: replace lighttpd with uhttpd
The default OpenWrt HTTP dameon is uhttpd, which supports all features
required by Facebook Wi-Fi.

Move the web files to the cgi-bin subdirectory, and add a json_script to
rewrite the URLs in the spec to the files in the cgi-bin subdirectory,
so that we can reuse the default uhttpd instance without changing the
cgi_prefix option. This way, LuCI keeps working.

Using a separate uhttpd instance for Facebook Wi-Fi will not work, as
the default uhttpd instance binds ports 80 and 443 already, which makes
it impossible for the next instance to do so.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>