fbwifi: flush chains before adding rules
Every time fbwifi_get_config is executed, the same rules are added to
both the FBWIFI_HOST_REDIRLIST and FBWIFI_TRAFFIC_ALLOWLIST iptables
chains. Due to this, they keep growing until manually flushed, or until
the device is rebooted. In case the data returned by the API changes,
the old rules will also never be removed.

Flush both chains before adding rules to avoid this. There will be a
tiny window in which no rules will exist, but this will not allow
clients to access the Internet before authenticating. It is possible to
avoid this, but it will require a lot of extra code and iptables calls,
so this is a reasonable trade-off.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>