fbwifi: add hotplug script
If the fbwifi service starts before connectivity is established, the
fbwifi_get_config script cannot get the config, and it will not retry,
leaving Facebook Wi-Fi in a broken state.

Add a hotplug script that will restart the service when an iface hotplug
event with ifup action for the WAN interface is detected to fix this.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>