fbwifi: introduce wrapper script
Introduce a wrapper script that will be used for changing UCI settings.
Doing this inside an init script, followed by reload_config, is
dangerous, as this will trigger init scripts, so we could end up in an
endless loop. Since this wrapper script will be used for changing UCI
settings, drop the UCI defaults and move them into the wrapper script.

As this wrapper script will modify the UCI firewall config, the
reload_config will also reload the firewall, and we can drop the
firewall restart from the init script.

Signed-off-by: John Crispin <john@phrozen.org>
Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>