fbwifi: use procd respawn instead of cron job
Instead of using crontab, we can simply use the procd respawn feature.
As the fbwifi_validate_token_db script currently restarts the service,
and that script is now actually the service, we could end up in an
endless loop. Move the reload handling to the wrapper script to avoid
that.

Signed-off-by: John Crispin <john@phrozen.org>
Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>