fbwifi: open https port in FBWIFI_INPUT
If the OpenWrt firewall config does not allow input traffic on the zone
used for Facebook Wi-Fi, clients will not be able to communicate with
the http service. Due to this, they will not be able to authenticate on
the captive portal. Allow https in the FBWIFI_INPUT chain to avoid this.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>