fbwifi: handle duplicate arp entries
Depending on the network config, a mac address can appear more than
once in /proc/net/arp. The current code doesn't handle that, and results
in various problems. To reduce the chance of that happening, match only
the lines with flag 0x2, which means this entry is reachable.
On underlying devices the flag would be 0x6 and when the entry is
expired it would be 0x0.

This should be solved in a cleaner way, but let's start with a quick fix
as this is currently breaking Facebook Wi-Fi on Open Wi-Fi.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>