fbwifi: set webroot for redirect uhttpd instance
When the redirect json script fails, uhttpd serves a directory listing
of its home directory. By default, this is the working directory, which
is the root filesystem. As a result, uhttpd will serve a directory
listing for the root filesystem, which is a security hazard.

Set the webroot to /dev/null to avoid this.

Reported-by: Mario Flajslik <mfl@fb.com>
Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>