fbwifi: rework UCI state tracking
Using UCI without commit will result in the state file growing forever.
A delete will result in a new line in the state file prefixed with a
hyphen, it will not delete the line from the state file.

To solve the problem of the growing state file, create the UCI cursor
with a custom confdir and savedir, and start using commit, so that the
changes in the state file will be saved in the config file.

To make this work, we need to make sure the config file exists,
otherwise the Lua UCI calls will silently fail.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>