fbwifi: fix CORS for info
The cacheloader module is quite complex, and messes with scoping in some
circumstances. In this case, when cacheloader is enabled, calling the
fbwifi module seems to overwrite the state variable. As a result, no
CORS headers are being set and requesting the info page returns 502.

As we don't know the exact implications of using the cacheloader module,
and it's unlikely to be beneficial for improving performance of just a
few lines of code, drop the use of the cacheloader module so that the
CORS headers are properly set.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>